requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
numpy>=1.24.0,<2.0.0
transformers==4.37.0
torch==2.1.2
accelerate==0.26.1
huggingface_hub==0.20.2
sentencepiece==0.1.99
protobuf==4.25.2
psycopg2-binary==2.9.9
boto3==1.34.0
sentence-transformers==2.2.2